Test: Small Simplification

+++ components/ILIAS/TestQuestionPool/classes/class.assFileUpload.php
@@ -334,7 +334,7 @@ class assFileUpload extends assQuestion implements ilObjQuestionScoringAdjustabl
             }
         }
 
-        return (float)0;
+        return 0.0;
     }
 
     protected function calculateReachedPointsForSolution($userSolution)
